+++ CHelpLib/Tests/tLinkedList_Perf.cpp
@@ -43,7 +43,7 @@ public:
 
         Helpers::ITimer* pTimer = new Helpers::CTimerTicks();
         pTimer->Start();
-        while ((nFieldsRead = fwscanf_s(pFile, L"%511s", szLine, ARRAYSIZE(szLine))) == 1)
+        while ((nFieldsRead = fwscanf_s(pFile, L"%511s", szLine, (UINT)ARRAYSIZE(szLine))) == 1)
         {
             // Skip strings that don't meet min length
             auto cch = wcslen(szLine);
